requirements.txt
appnope==0.1.4
asttokens==3.0.0
boto3==1.40.11
botocore==1.40.11
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
comm==0.2.3
DateTime==5.5
debugpy==1.8.16
decorator==5.2.1
et_xmlfile==2.0.0
executing==2.2.0
geopandas==1.1.1
google-auth==2.40.3
google-auth-oauthlib==1.2.2
gspread==6.2.1
idna==3.10
ipykernel==6.30.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jmespath==1.0.1
jupyter_client==8.6.3
jupyter_core==5.8.1
logging==0.4.9.6
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.3.2
oauthlib==3.3.1
openpyxl==3.1.5
ordereddict==1.1
packaging==25.0
pandas==2.3.1
parso==0.8.4
pathlib==1.0.1
pexpect==4.9.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
Pygments==2.19.2
pyogrio==0.11.1
pyproj==3.7.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.1
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
s3transfer==0.13.1
setuptools==80.9.0
shapely==2.1.1
six==1.17.0
stack-data==0.6.3
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
xlsxwriter==3.2.5
zope.interface==7.2
